○ver1.0.17 (2018/11/11)
・パーツ選択画面で強化したパーツの積載猶予が情報なしになる不具合を修正。
・アセン画面や性能画面の強化済みパーツの重量が総重量に加算されないことがある不具合を修正。
・UK57速射機銃・弐の威力(強化3時)のデータ誤りを修正。

+++ app/src/main/java/hokage/kaede/gmail/com/BBViewLib/Java/CustomData.java
@@ -1041,8 +1041,8 @@ public class CustomData {
 	 * @param blust_type 総重量を取得する兵装
 	 * @return 総重量
 	 */
-	private int getWeaponsWeight(String blust_type) {
-		int ret = 0;
+	private double getWeaponsWeight(String blust_type) {
+		double ret = 0;
 		
 		BBData[] blust_data = getWeaponList(blust_type);
 		int blust_len = blust_data.length;
@@ -1050,7 +1050,7 @@ public class CustomData {
 		try {
 			for(int i=0; i<blust_len; i++) {
 				String str_buf = blust_data[i].get("重量");
-				int buf = Integer.parseInt(str_buf);
+				double buf = Double.parseDouble(str_buf);
 				ret = ret + buf;
 			}
 		} catch(Exception e) {
